Adds a common history and screen between screen types
