+++ src/ui/mod.rs
@@ -3,7 +3,8 @@ pub use self::{
     command::spawn_input_thread,
     help_handler::HelpHandler,
     reader_screen::ReaderScreen,
-    screen::Screen,
+    screen::{create_screen, create_screen_writer, switch_screen},
+    split_screen::SplitScreen,
     user_interface::{wrap_line, UserInterface},
 };
 
@@ -13,6 +14,9 @@ pub use self::user_interface::MockUserInterface;
 mod ansi;
 mod command;
 mod help_handler;
+mod history;
 mod reader_screen;
 mod screen;
+mod scroll_data;
+mod split_screen;
 mod user_interface;
